Create User function
Created a new window in Flightbooking.UI that can be used to create a new user. The code behind the UI sends the information entered to the BLL where a very basic validation takes place then it is sent to the DAL layer which in turn Inserts the Username and Password into the user database. Has been tested to work but the button that opens the create user window is proportioned weirdly short.